Updated to MC 1.21.6

+++ src/main/java/com/g4mesoft/ui/renderer/GSClipRect.java
@@ -2,19 +2,19 @@ package com.g4mesoft.ui.renderer;
 
 public class GSClipRect {
 
-	public static final GSClipRect EMPTY = new GSClipRect(0.0f, 0.0f, 0.0f, 0.0f);
+	public static final GSClipRect EMPTY = new GSClipRect(0, 0, 0, 0);
 
-	public final float x0;
-	public final float y0;
+	public final int x0;
+	public final int y0;
 	
-	public final float x1;
-	public final float y1;
+	public final int x1;
+	public final int y1;
 	
 	public GSClipRect(GSClipRect bounds) {
 		this(bounds.x0, bounds.y0, bounds.x1, bounds.y1);
 	}
 
-	public GSClipRect(float x0, float y0, float x1, float y1) {
+	public GSClipRect(int x0, int y0, int x1, int y1) {
 		this.x0 = x0;
 		this.y0 = y0;
 		
@@ -22,16 +22,16 @@ public class GSClipRect {
 		this.y1 = y1;
 	}
 
-	public GSClipRect offset(float clipOffsetX, float clipOffsetY) {
+	public GSClipRect offset(int clipOffsetX, int clipOffsetY) {
 		return new GSClipRect(x0 + clipOffsetX, y0 + clipOffsetY, 
 		                      x1 + clipOffsetX, y1 + clipOffsetY);
 	}
 
 	public GSClipRect intersection(GSClipRect other) {
-		float _x0 = Math.max(x0, other.x0);
-		float _y0 = Math.max(y0, other.y0);
-		float _x1 = Math.min(x1, other.x1);
-		float _y1 = Math.min(y1, other.y1);
+		int _x0 = Math.max(x0, other.x0);
+		int _y0 = Math.max(y0, other.y0);
+		int _x1 = Math.min(x1, other.x1);
+		int _y1 = Math.min(y1, other.y1);
 		if (_x0 >= _x1 || _y0 >= _y1) {
 			return GSClipRect.EMPTY;
 		} else {
@@ -40,10 +40,10 @@ public class GSClipRect {
 	}
 
 	public GSClipRect union(GSClipRect other) {
-		float _x0 = Math.min(x0, other.x0);
-		float _y0 = Math.min(y0, other.y0);
-		float _x1 = Math.max(x1, other.x1);
-		float _y1 = Math.max(y1, other.y1);
+		int _x0 = Math.min(x0, other.x0);
+		int _y0 = Math.min(y0, other.y0);
+		int _x1 = Math.max(x1, other.x1);
+		int _y1 = Math.max(y1, other.y1);
 		return new GSClipRect(_x0, _y0, _x1, _y1);
 	}
 }
